Update tour.md
I added a part 1 that is just describing the key value-add conceptually without the detailed example. This is for readers to quickly get the key value-adds without getting distracted by detailed code snippets, which are quite long. 

part 2 will be code snippets. which need to be updated. and I will